refactor(xmlrpc): split client into per-resource modules
XmlRpcClient mixed transport, per-resource request methods, and response
mappers for bug/comment/attachment/user/group in one 655-line file, while the
REST client is already split per resource. Extract:

- client.rs keeps only transport (struct, new, call); call is now pub(crate)
- mappers.rs holds the shared value-access primitives and lookup_bug_entry
- bug.rs / comment.rs / attachment.rs / user.rs / group.rs each own their
  resource methods and mappers via impl XmlRpcClient blocks

Tests move to per-module sibling files; the 110 xmlrpc tests are preserved.
No behavior or public-API change — XmlRpcClient is reachable exactly as before.
