+++ src/xmlrpc/comment.rs
@@ -1,0 +1,88 @@
+use std::collections::BTreeMap;
+
+use crate::error::{BzrError, Result};
+use crate::xmlrpc::client::XmlRpcClient;
+use crate::xmlrpc::mappers::{
+    get_bool_flag, get_datetime_str, get_nonempty_str, get_str, lookup_bug_entry,
+};
+use crate::xmlrpc::value::Value;
+
+impl XmlRpcClient {
+    pub async fn get_comments_since(
+        &self,
+        bug_id: u64,
+        since: Option<&str>,
+    ) -> Result<Vec<crate::types::Comment>> {
+        let mut rpc_params = BTreeMap::new();
+        #[expect(clippy::cast_possible_wrap, reason = "bug IDs fit in i64")]
+        let bug_id_value = Value::Int(bug_id as i64);
+        rpc_params.insert("ids".into(), Value::Array(vec![bug_id_value]));
+        if let Some(s) = since {
+            rpc_params.insert("new_since".into(), Value::from(s));
+        }
+
+        let result = self.call("Bug.comments", rpc_params).await?;
+        extract_comments(&result, bug_id)
+    }
+}
+
+fn extract_comments(response: &Value, bug_id: u64) -> Result<Vec<crate::types::Comment>> {
+    let Some(bug_entry) = lookup_bug_entry(response, bug_id)? else {
+        return Ok(Vec::new());
+    };
+
+    let entry_struct = bug_entry
+        .as_struct()
+        .ok_or_else(|| BzrError::XmlRpc("expected bug entry struct".into()))?;
+
+    let Some(comments_val) = entry_struct.get("comments") else {
+        return Ok(Vec::new());
+    };
+
+    let comments_arr = comments_val
+        .as_array()
+        .ok_or_else(|| BzrError::XmlRpc("expected comments array".into()))?;
+
+    let mut comments = Vec::with_capacity(comments_arr.len());
+    for c in comments_arr {
+        comments.push(value_to_comment(c)?);
+    }
+    Ok(comments)
+}
+
+fn value_to_comment(val: &Value) -> Result<crate::types::Comment> {
+    let m = val
+        .as_struct()
+        .ok_or_else(|| BzrError::XmlRpc("expected struct for comment".into()))?;
+
+    let id = m
+        .get("id")
+        .and_then(Value::as_i64)
+        .ok_or_else(|| BzrError::XmlRpc("comment missing id field".into()))?;
+    let bug_id = m.get("bug_id").and_then(Value::as_i64).unwrap_or(0);
+    let count = m.get("count").and_then(Value::as_i64).unwrap_or(0);
+
+    #[expect(clippy::cast_sign_loss, reason = "comment IDs are non-negative")]
+    let id = id as u64;
+    #[expect(clippy::cast_sign_loss, reason = "bug IDs are non-negative")]
+    let bug_id = bug_id as u64;
+    #[expect(clippy::cast_sign_loss, reason = "comment counts are non-negative")]
+    let count = count as u64;
+    Ok(crate::types::Comment {
+        id,
+        bug_id,
+        text: get_str(m, "text").unwrap_or_default(),
+        creator: get_nonempty_str(m, "creator"),
+        creation_time: get_datetime_str(m, "creation_time"),
+        count,
+        is_private: get_bool_flag(m, "is_private"),
+        attachment_id: m
+            .get("attachment_id")
+            .and_then(Value::as_i64)
+            .and_then(|v| u64::try_from(v).ok()),
+    })
+}
+
+#[cfg(test)]
+#[path = "comment_tests.rs"]
+mod tests;
